pyproject.toml
[project]
name = "qm-qua"
description = "QUA language SDK to control a Quantum Computer"
readme = "README.md"
authors = [{ name = "Quantum Machines", email = "info@quantum-machines.co"}]
license = "BSD-3-Clause"
dynamic = [ "classifiers", "version" ]

[project.urls]
Homepage = "https://www.quantum-machines.co/"
Documentation = "https://docs.quantum-machines.co/"
Repository = "https://github.com/qm-labs/qm-qua-sdk-public"
Download = "https://pypi.org/project/qm-qua/#files"
Changelog = "https://github.com/qm-labs/qm-qua-sdk-public/blob/main/CHANGELOG.md"

[tool.poetry]
name = "qm-qua"
version = "0.0.0"
description = "QUA language SDK to control a Quantum Computer"
authors = ["Quantum Machines <info@quantum-machines.co>"]
classifiers = ["Topic :: Scientific/Engineering :: Quantum Computing", "Development Status :: 5 - Production/Stable"]

packages = [
    { include = "qm" },
    { include = "qm/py.typed" }
]
include = [
    "CHANGELOG.md",
    "LICENSE",
    {path="qm/io/**/*", format=["sdist", "wheel"]},
    {path="qm/grpc/**/*.py", format=["sdist", "wheel"]},
    {path="qm/grpc/**/*.pyi", format=["sdist", "wheel"]},
    {path="qm/grpc/VERSION", format=["sdist", "wheel"]},
]


[tool.poetry.dependencies]
python = ">=3.9, <3.14"
grpcio = "^1.57"
protobuf = ">=4.25, <7"
marshmallow = ">=3,<4"
marshmallow-polyfield = "^5.7"
numpy = "<3"
tinydb = "^4.6.1"
certifi = { version = "*", optional = true }
datadog-api-client = "^2.6.0"
typing-extensions = "^4.5"

plotly = ">=5"
httpx = { version = ">=0.27.0,<1", extras = ["http2"] }  # Note: httpx 0.28.0+ has reported issues in Jupyter notebooks. Make sure to test if considering further version restrictions.
multidict = "^6.7.0"

[tool.poetry.group.dev.dependencies]
boto3 = "^1.42.30"
poethepoet = "^0.16.0"
grpcio-tools = "==1.68.1"
pytest-asyncio = "^0.20.3"
pytest-benchmark = { extras = ["aspect"], version = "^4.0.0" }
pytest = "^7.0.0"
pytest-cov = "^4.0.0"
pytest-order = "^1.0.0"
pytest-xdist = "^3.8"
poetry ="<=2.1.3"
coverage = "^7.0.1"
black = "^22.6.0"
flake8 = "^3.9.2"
flake8-bugbear = "^21.11.29"

apispec = "^5.1.1"
poetry-dynamic-versioning = { extras = ["plugin"], version = "^1.0.0" }
mypy = "^1.14"
isort = "^5.11.5"
qm-saas = "^1.1.7"

[tool.poetry.extras]
simulation = ["certifi"]

[tool.black]
line-length = 120
target-version = ["py39"]
extend-exclude = '''
/(
  | qm/grpc/.*
  | qm/io/.*
)/
'''

[tool.isort]
profile = "black"
line_length = 120
length_sort = true
skip_gitignore = true
force_sort_within_sections = true
extra_standard_library = ["typing_extensions"]


[tool.poe.tasks]

clean-proto.sequence = [
    { script = "shutil:rmtree('qm/proto', ignore_errors=True)" },
]

clean-grpc.sequence = [
    { script = "shutil:rmtree('qm/grpc', ignore_errors=True)" },
    { script = "shutil:rmtree('qm/pb', ignore_errors=True)" },
    { script = "shutil:rmtree('qm/io', ignore_errors=True)" },
    { script = "os:mkdir('qm/grpc')" },
]

download-proto = "python scripts/download_proto.py"

create-grpc.shell = "bash scripts/generate_grpc.sh $PROTO_SRC_PATH"
create-grpc.args = [
  { name = "PROTO_SRC_PATH", positional = true, default = "." }
]

copy-version-grpc = "cp proto/VERSION qm/grpc/VERSION"
postprocess-protobuf = "python scripts/postprocess_proto_generation.py"

generate-grpc-from-dir.sequence = [
    "clean-grpc",
    "create-grpc $PROTO_SRC_PATH",
    "copy-version-grpc",
    "postprocess-protobuf"
]
generate-grpc-from-dir.args = [
  { name = "PROTO_SRC_PATH", positional = true, default = "./" }
]


generate-grpc.sequence = [
    "clean-proto",
    "download-proto",
    "generate-grpc-from-dir",
]

check-grpc-breaking-changes = "buf breaking --against '.git#branch=main'"

format-black = "poetry run black qm tests"
format-isort = "poetry run isort qm tests"
format.sequence = ["format-black", "format-isort"]
format.ignore_fail = true

lint-black = "black qm tests --check"  # Also checks the tests
lint-flake8 = "flake8 qm"
lint-mypy = "mypy qm"
lint.sequence = ["lint-black", "lint-flake8", "lint-mypy"]
lint.ignore_fail = true

test = "pytest tests"

generate-dsl-shortcut = "poetry run python for_docs/scripts/generate_qua_dsl_shortcut.py"
build-config-schema = "poetry run python for_docs/scripts/build_config_schema.py"

generate-files-for-docs-website.sequence = ["build-config-schema", "generate-dsl-shortcut"]

check.sequence = ["lint", "test"]
check.ignore_fail = true

pre-commit-hook.sequence = ["format", "lint", "generate-files-for-docs-website"]
pre-commit-hook.ignore_fail = true


[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true

[tool.poetry-dynamic-versioning]
enable = true
format-jinja = "{{ env['BASE_VERSION'] | default(base) }}{{ env['VERSION_SUFFIX'] | default('') }}"

[tool.poetry-dynamic-versioning.substitution]
files = ["qm/version.py"]
patterns = ["(^__version__\\s*(?::.*?)?=\\s*['\"])[^'\"]*(['\"])"]

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning>=1.0.0,<2.0.0"]
build-backend = "poetry_dynamic_versioning.backend"
